+++ playbilling/src/main/java/com/google/androidbrowserhelper/playbilling/provider/BillingWrapper.java
@@ -21,7 +21,6 @@ import com.android.billingclient.api.BillingClient;
 import com.android.billingclient.api.BillingClientStateListener;
 import com.android.billingclient.api.BillingResult;
 import com.android.billingclient.api.ConsumeResponseListener;
-import com.android.billingclient.api.PriceChangeConfirmationListener;
 import com.android.billingclient.api.PurchaseHistoryResponseListener;
 import com.android.billingclient.api.PurchasesResponseListener;
 import com.android.billingclient.api.SkuDetails;
@@ -78,11 +77,5 @@ public interface BillingWrapper {
      * {@link Listener#onPurchaseFlowComplete} should be called.
      */
     boolean launchPaymentFlow(Activity activity, SkuDetails sku, MethodData methodData);
-
-    /**
-     * Launches the price change confirmation flow.
-     */
-    void launchPriceChangeConfirmationFlow(Activity activity, SkuDetails sku,
-            PriceChangeConfirmationListener listener);
 
 }